+++ include/prtree/core/prtree.h
@@ -1,4 +1,6 @@
 #pragma once
+
+// Standard Library Includes
 #include <algorithm>
 #include <array>
 #include <atomic>
@@ -22,9 +24,11 @@
 #include <unordered_map>
 #include <utility>
 #include <vector>
-// Phase 8: C++20 features
+
+// C++20 features
 #include <concepts>
 
+// External Dependencies
 #include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
@@ -34,14 +38,22 @@
 #include <cereal/cereal.hpp>
 #include <cereal/types/array.hpp>
 #include <cereal/types/atomic.hpp>
-#include <cereal/types/memory.hpp> //for smart pointers
+#include <cereal/types/memory.hpp>
 #include <cereal/types/string.hpp>
 #include <cereal/types/unordered_map.hpp>
 #include <cereal/types/vector.hpp>
 
+#include <snappy.h>
+
+// PRTree Modular Components
+#include "prtree/core/detail/types.h"
+#include "prtree/core/detail/bounding_box.h"
+#include "prtree/core/detail/data_type.h"
+#include "prtree/core/detail/pseudo_tree.h"
+#include "prtree/core/detail/nodes.h"
+
 #include "prtree/utils/parallel.h"
 #include "prtree/utils/small_vector.h"
-#include <snappy.h>
 
 #ifdef MY_DEBUG
 #include <gperftools/profiler.h>
@@ -49,596 +61,8 @@
 
 using Real = float;
 
-// Phase 4: Versioning for serialization
-constexpr uint16_t PRTREE_VERSION_MAJOR = 1;
-constexpr uint16_t PRTREE_VERSION_MINOR = 0;
-
 namespace py = pybind11;
 
-// Phase 8: C++20 Concepts for type safety
-template <typename T>
-concept IndexType = std::integral<T> && !std::same_as<T, bool>;
-
-template <typename T>
-concept SignedIndexType = IndexType<T> && std::is_signed_v<T>;
-
-template <class T> using vec = std::vector<T>;
-
-template <typename Sequence>
-inline py::array_t<typename Sequence::value_type> as_pyarray(Sequence &seq) {
-
-  auto size = seq.size();
-  auto data = seq.data();
-  std::unique_ptr<Sequence> seq_ptr =
-      std::make_unique<Sequence>(std::move(seq));
-  auto capsule = py::capsule(seq_ptr.get(), [](void *p) {
-    std::unique_ptr<Sequence>(reinterpret_cast<Sequence *>(p));
-  });
-  seq_ptr.release();
-  return py::array(size, data, capsule);
-}
-
-template <typename T> auto list_list_to_arrays(vec<vec<T>> out_ll) {
-  vec<T> out_s;
-  out_s.reserve(out_ll.size());
-  std::size_t sum = 0;
-  for (auto &&i : out_ll) {
-    out_s.push_back(i.size());
-    sum += i.size();
-  }
-  vec<T> out;
-  out.reserve(sum);
-  for (const auto &v : out_ll)
-    out.insert(out.end(), v.begin(), v.end());
-
-  return make_tuple(std::move(as_pyarray(out_s)), std::move(as_pyarray(out)));
-}
-
-template <class T, size_t StaticCapacity>
-using svec = itlib::small_vector<T, StaticCapacity>;
-
-template <class T> using deque = std::deque<T>;
-
-template <class T> using queue = std::queue<T, deque<T>>;
-
-static const float REBUILD_THRE = 1.25;
-
-// Phase 8: Branch prediction hints
-// Note: C++20 provides [[likely]] and [[unlikely]] attributes, but we keep
-// these macros for backward compatibility and cleaner syntax in conditions.
-// Future refactoring could replace: if (unlikely(x)) with if (x) [[unlikely]]
-#if defined(__GNUC__) || defined(__clang__)
-#define likely(x) __builtin_expect(!!(x), 1)
-#define unlikely(x) __builtin_expect(!!(x), 0)
-#else
-#define likely(x) (x)
-#define unlikely(x) (x)
-#endif
-
-std::string compress(std::string &data) {
-  std::string output;
-  snappy::Compress(data.data(), data.size(), &output);
-  return output;
-}
-
-std::string decompress(std::string &data) {
-  std::string output;
-  snappy::Uncompress(data.data(), data.size(), &output);
-  return output;
-}
-
-template <int D = 2> class BB {
-private:
-  Real values[2 * D];
-
-public:
-  BB() { clear(); }
-
-  BB(const Real (&minima)[D], const Real (&maxima)[D]) {
-    Real v[2 * D];
-    for (int i = 0; i < D; ++i) {
-      v[i] = -minima[i];
-      v[i + D] = maxima[i];
-    }
-    validate(v);
-    for (int i = 0; i < D; ++i) {
-      values[i] = v[i];
-      values[i + D] = v[i + D];
-    }
-  }
-
-  BB(const Real (&v)[2 * D]) {
-    validate(v);
-    for (int i = 0; i < D; ++i) {
-      values[i] = v[i];
-      values[i + D] = v[i + D];
-    }
-  }
-
-  Real min(const int dim) const {
-    if (unlikely(dim < 0 || D <= dim)) {
-      throw std::runtime_error("Invalid dim");
-    }
-    return -values[dim];
-  }
-  Real max(const int dim) const {
-    if (unlikely(dim < 0 || D <= dim)) {
-      throw std::runtime_error("Invalid dim");
-    }
-    return values[dim + D];
-  }
-
-  bool validate(const Real (&v)[2 * D]) const {
-    bool flag = false;
-    for (int i = 0; i < D; ++i) {
-      if (unlikely(-v[i] > v[i + D])) {
-        flag = true;
-        break;
-      }
-    }
-    if (unlikely(flag)) {
-      throw std::runtime_error("Invalid Bounding Box");
-    }
-    return flag;
-  }
-  void clear() noexcept {
-    for (int i = 0; i < 2 * D; ++i) {
-      values[i] = -1e100;
-    }
-  }
-
-  Real val_for_comp(const int &axis) const noexcept {
-    const int axis2 = (axis + 1) % (2 * D);
-    return values[axis] + values[axis2];
-  }
-
-  BB operator+(const BB &rhs) const {
-    Real result[2 * D];
-    for (int i = 0; i < 2 * D; ++i) {
-      result[i] = std::max(values[i], rhs.values[i]);
-    }
-    return BB<D>(result);
-  }
-
-  BB operator+=(const BB &rhs) {
-    for (int i = 0; i < 2 * D; ++i) {
-      values[i] = std::max(values[i], rhs.values[i]);
-    }
-    return *this;
-  }
-
-  void expand(const Real (&delta)[D]) noexcept {
-    for (int i = 0; i < D; ++i) {
-      values[i] += delta[i];
-      values[i + D] += delta[i];
-    }
-  }
-
-  bool operator()(
-      const BB &target) const { // whether this and target has any intersect
-
-    Real minima[D];
-    Real maxima[D];
-    bool flags[D];
-    bool flag = true;
-
-    for (int i = 0; i < D; ++i) {
-      minima[i] = std::min(values[i], target.values[i]);
-      maxima[i] = std::min(values[i + D], target.values[i + D]);
-    }
-    for (int i = 0; i < D; ++i) {
-      flags[i] = -minima[i] <= maxima[i];
-    }
-    for (int i = 0; i < D; ++i) {
-      flag &= flags[i];
-    }
-    return flag;
-  }
-
-  Real area() const {
-    Real result = 1;
-    for (int i = 0; i < D; ++i) {
-      result *= max(i) - min(i);
-    }
-    return result;
-  }
-
-  inline Real operator[](const int i) const { return values[i]; }
-
-  template <class Archive> void serialize(Archive &ar) { ar(values); }
-};
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int D = 2> class DataType {
-public:
-  BB<D> second;
-  T first;
-
-  DataType() noexcept = default;
-
-  DataType(const T &f, const BB<D> &s) {
-    first = f;
-    second = s;
-  }
-
-  DataType(T &&f, BB<D> &&s) noexcept {
-    first = std::move(f);
-    second = std::move(s);
-  }
-
-  void swap(DataType& other) noexcept {
-    using std::swap;
-    swap(first, other.first);
-    swap(second, other.second);
-  }
-
-  template <class Archive> void serialize(Archive &ar) { ar(first, second); }
-};
-
-template <class T, int D = 2>
-void clean_data(DataType<T, D> *b, DataType<T, D> *e) {
-  for (DataType<T, D> *it = e - 1; it >= b; --it) {
-    it->~DataType<T, D>();
-  }
-}
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int B = 6, int D = 2> class Leaf {
-public:
-  BB<D> mbb;
-  svec<DataType<T, D>, B> data; // You can swap when filtering
-  int axis = 0;
-
-  // T is type of keys(ids) which will be returned when you post a query.
-  Leaf() { mbb = BB<D>(); }
-  Leaf(const int _axis) {
-    axis = _axis;
-    mbb = BB<D>();
-  }
-
-  void set_axis(const int &_axis) { axis = _axis; }
-
-  void push(const T &key, const BB<D> &target) {
-    data.emplace_back(key, target);
-    update_mbb();
-  }
-
-  void update_mbb() {
-    mbb.clear();
-    for (const auto &datum : data) {
-      mbb += datum.second;
-    }
-  }
-
-  bool filter(DataType<T, D> &value) { // false means given value is ignored
-    // Phase 2: C++20 requires explicit 'this' capture
-    auto comp = [this](const auto &a, const auto &b) noexcept {
-      return a.second.val_for_comp(axis) < b.second.val_for_comp(axis);
-    };
-
-    if (data.size() < B) { // if there is room, just push the candidate
-      auto iter = std::lower_bound(data.begin(), data.end(), value, comp);
-      DataType<T, D> tmp_value = DataType<T, D>(value);
-      data.insert(iter, std::move(tmp_value));
-      mbb += value.second;
-      return true;
-    } else { // if there is no room, check the priority and swap if needed
-      if (data[0].second.val_for_comp(axis) < value.second.val_for_comp(axis)) {
-        size_t n_swap =
-            std::lower_bound(data.begin(), data.end(), value, comp) -
-            data.begin();
-        std::swap(*data.begin(), value);
-        auto iter = data.begin();
-        for (size_t i = 0; i < n_swap - 1; ++i) {
-          std::swap(*(iter + i), *(iter + i + 1));
-        }
-        update_mbb();
-      }
-      return false;
-    }
-  }
-};
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int B = 6, int D = 2> class PseudoPRTreeNode {
-public:
-  Leaf<T, B, D> leaves[2 * D];
-  std::unique_ptr<PseudoPRTreeNode> left, right;
-
-  PseudoPRTreeNode() {
-    for (int i = 0; i < 2 * D; i++) {
-      leaves[i].set_axis(i);
-    }
-  }
-  PseudoPRTreeNode(const int axis) {
-    for (int i = 0; i < 2 * D; i++) {
-      const int j = (axis + i) % (2 * D);
-      leaves[i].set_axis(j);
-    }
-  }
-
-  template <class Archive> void serialize(Archive &archive) {
-    // archive(cereal::(left), cereal::defer(right), leaves);
-    archive(left, right, leaves);
-  }
-
-  void address_of_leaves(vec<Leaf<T, B, D> *> &out) {
-    for (auto &leaf : leaves) {
-      if (leaf.data.size() > 0) {
-        out.emplace_back(&leaf);
-      }
-    }
-  }
-
-  template <class iterator> auto filter(const iterator &b, const iterator &e) {
-    auto out = std::remove_if(b, e, [&](auto &x) {
-      for (auto &l : leaves) {
-        if (l.filter(x)) {
-          return true;
-        }
-      }
-      return false;
-    });
-    return out;
-  }
-};
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int B = 6, int D = 2> class PseudoPRTree {
-public:
-  std::unique_ptr<PseudoPRTreeNode<T, B, D>> root;
-  vec<Leaf<T, B, D> *> cache_children;
-  const int nthreads = std::max(1, (int)std::thread::hardware_concurrency());
-
-  PseudoPRTree() { root = std::make_unique<PseudoPRTreeNode<T, B, D>>(); }
-
-  template <class iterator> PseudoPRTree(const iterator &b, const iterator &e) {
-    if (!root) {
-      root = std::make_unique<PseudoPRTreeNode<T, B, D>>();
-    }
-    construct(root.get(), b, e, 0);
-    clean_data<T, D>(b, e);
-  }
-
-  template <class Archive> void serialize(Archive &archive) {
-    archive(root);
-    // archive.serializeDeferments();
-  }
-
-  template <class iterator>
-  void construct(PseudoPRTreeNode<T, B, D> *node, const iterator &b,
-                 const iterator &e, const int depth) {
-    if (e - b > 0 && node != nullptr) {
-      bool use_recursive_threads = std::pow(2, depth + 1) <= nthreads;
-#ifdef MY_DEBUG
-      use_recursive_threads = false;
-#endif
-
-      vec<std::thread> threads;
-      threads.reserve(2);
-      PseudoPRTreeNode<T, B, D> *node_left, *node_right;
-
-      const int axis = depth % (2 * D);
-      auto ee = node->filter(b, e);
-      auto m = b;
-      std::advance(m, (ee - b) / 2);
-      std::nth_element(b, m, ee,
-                       [axis](const DataType<T, D> &lhs,
-                              const DataType<T, D> &rhs) noexcept {
-                         return lhs.second[axis] < rhs.second[axis];
-                       });
-
-      if (m - b > 0) {
-        node->left = std::make_unique<PseudoPRTreeNode<T, B, D>>(axis);
-        node_left = node->left.get();
-        if (use_recursive_threads) {
-          threads.push_back(
-              std::thread([&]() { construct(node_left, b, m, depth + 1); }));
-        } else {
-          construct(node_left, b, m, depth + 1);
-        }
-      }
-      if (ee - m > 0) {
-        node->right = std::make_unique<PseudoPRTreeNode<T, B, D>>(axis);
-        node_right = node->right.get();
-        if (use_recursive_threads) {
-          threads.push_back(
-              std::thread([&]() { construct(node_right, m, ee, depth + 1); }));
-        } else {
-          construct(node_right, m, ee, depth + 1);
-        }
-      }
-      std::for_each(threads.begin(), threads.end(),
-                    [&](std::thread &x) { x.join(); });
-    }
-  }
-
-  auto get_all_leaves(const int hint) {
-    if (cache_children.empty()) {
-      using U = PseudoPRTreeNode<T, B, D>;
-      cache_children.reserve(hint);
-      auto node = root.get();
-      queue<U *> que;
-      que.emplace(node);
-
-      while (!que.empty()) {
-        node = que.front();
-        que.pop();
-        node->address_of_leaves(cache_children);
-        if (node->left)
-          que.emplace(node->left.get());
-        if (node->right)
-          que.emplace(node->right.get());
-      }
-    }
-    return cache_children;
-  }
-
-  std::pair<DataType<T, D> *, DataType<T, D> *> as_X(void *placement,
-                                                     const int hint) {
-    DataType<T, D> *b, *e;
-    auto children = get_all_leaves(hint);
-    T total = children.size();
-    b = reinterpret_cast<DataType<T, D> *>(placement);
-    e = b + total;
-    for (T i = 0; i < total; i++) {
-      new (b + i) DataType<T, D>{i, children[i]->mbb};
-    }
-    return {b, e};
-  }
-};
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int B = 6, int D = 2> class PRTreeLeaf {
-public:
-  BB<D> mbb;
-  svec<DataType<T, D>, B> data;
-
-  PRTreeLeaf() { mbb = BB<D>(); }
-
-  PRTreeLeaf(const Leaf<T, B, D> &leaf) {
-    mbb = leaf.mbb;
-    data = leaf.data;
-  }
-
-  Real area() const { return mbb.area(); }
-
-  void update_mbb() {
-    mbb.clear();
-    for (const auto &datum : data) {
-      mbb += datum.second;
-    }
-  }
-
-  void operator()(const BB<D> &target, vec<T> &out) const {
-    if (mbb(target)) {
-      for (const auto &x : data) {
-        if (x.second(target)) {
-          out.emplace_back(x.first);
-        }
-      }
-    }
-  }
-
-  void del(const T &key, const BB<D> &target) {
-    if (mbb(target)) {
-      auto remove_it =
-          std::remove_if(data.begin(), data.end(), [&](auto &datum) {
-            return datum.second(target) && datum.first == key;
-          });
-      data.erase(remove_it, data.end());
-    }
-  }
-
-  void push(const T &key, const BB<D> &target) {
-    data.emplace_back(key, target);
-    update_mbb();
-  }
-
-  template <class Archive> void save(Archive &ar) const {
-    vec<DataType<T, D>> _data;
-    for (const auto &datum : data) {
-      _data.push_back(datum);
-    }
-    ar(mbb, _data);
-  }
-
-  template <class Archive> void load(Archive &ar) {
-    vec<DataType<T, D>> _data;
-    ar(mbb, _data);
-    for (const auto &datum : _data) {
-      data.push_back(datum);
-    }
-  }
-};
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int B = 6, int D = 2> class PRTreeNode {
-public:
-  BB<D> mbb;
-  std::unique_ptr<Leaf<T, B, D>> leaf;
-  std::unique_ptr<PRTreeNode<T, B, D>> head, next;
-
-  PRTreeNode() {}
-  PRTreeNode(const BB<D> &_mbb) { mbb = _mbb; }
-
-  PRTreeNode(BB<D> &&_mbb) noexcept { mbb = std::move(_mbb); }
-
-  PRTreeNode(Leaf<T, B, D> *l) {
-    leaf = std::make_unique<Leaf<T, B, D>>();
-    mbb = l->mbb;
-    leaf->mbb = std::move(l->mbb);
-    leaf->data = std::move(l->data);
-  }
-
-  bool operator()(const BB<D> &target) { return mbb(target); }
-};
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int B = 6, int D = 2> class PRTreeElement {
-public:
-  BB<D> mbb;
-  std::unique_ptr<PRTreeLeaf<T, B, D>> leaf;
-  bool is_used = false;
-
-  PRTreeElement() {
-    mbb = BB<D>();
-    is_used = false;
-  }
-
-  PRTreeElement(const PRTreeNode<T, B, D> &node) {
-    mbb = BB<D>(node.mbb);
-    if (node.leaf) {
-      Leaf<T, B, D> tmp_leaf = Leaf<T, B, D>(*node.leaf.get());
-      leaf = std::make_unique<PRTreeLeaf<T, B, D>>(tmp_leaf);
-    }
-    is_used = true;
-  }
-
-  bool operator()(const BB<D> &target) { return is_used && mbb(target); }
-
-  template <class Archive> void serialize(Archive &archive) {
-    archive(mbb, leaf, is_used);
-  }
-};
-
-// Phase 8: Apply C++20 concept constraints
-template <IndexType T, int B = 6, int D = 2>
-void bfs(
-    const std::function<void(std::unique_ptr<PRTreeLeaf<T, B, D>> &)> &func,
-    vec<PRTreeElement<T, B, D>> &flat_tree, const BB<D> target) {
-  queue<size_t> que;
-  auto qpush_if_intersect = [&](const size_t &i) {
-    PRTreeElement<T, B, D> &r = flat_tree[i];
-    // std::cout << "i " << (long int) i << " : " << (bool) r.leaf << std::endl;
-    if (r(target)) {
-      // std::cout << " is pushed" << std::endl;
-      que.emplace(i);
-    }
-  };
-
-  // std::cout << "size: " << flat_tree.size() << std::endl;
-  qpush_if_intersect(0);
-  while (!que.empty()) {
-    size_t idx = que.front();
-    // std::cout << "idx: " << (long int) idx << std::endl;
-    que.pop();
-    PRTreeElement<T, B, D> &elem = flat_tree[idx];
-
-    if (elem.leaf) {
-      // std::cout << "func called for " << (long int) idx << std::endl;
-      func(elem.leaf);
-    } else {
-      for (size_t offset = 0; offset < B; offset++) {
-        size_t jdx = idx * B + offset + 1;
-        qpush_if_intersect(jdx);
-      }
-    }
-  }
-}
-
-// Phase 8: Apply C++20 concept constraints for type safety
-// T must be an integral type (used as index), not bool
 template <IndexType T, int B = 6, int D = 2> class PRTree {
 private:
   vec<PRTreeElement<T, B, D>> flat_tree;
